Refactor : move method(LadderGame class -> LadderGameApp)
